requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=dev.txt dev.in
#
argon2-cffi==19.1.0
backcall==0.1.0           # via ipython
certifi==2019.6.16        # via requests
cffi==1.12.3              # via argon2-cffi
chardet==3.0.4            # via requests
click==7.0                # via pip-tools
decorator==4.4.0          # via ipython, traitlets
django-cacheops==4.1
django-crispy-forms==1.7.2
django-environ==0.4.5
django-extensions==2.2.1
django-ipware==2.1.0
django-model-utils==3.2.0
django==2.2.5
djangorestframework==3.10.3
funcy==1.13               # via django-cacheops
geoip2==2.9.0
idna==2.8                 # via requests
ipdb==0.12.2
ipython-genutils==0.2.0   # via traitlets
ipython==7.7.0
jedi==0.15.1              # via ipython
maxminddb==1.4.1          # via geoip2
parso==0.5.1              # via jedi
pexpect==4.7.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==6.1.0
pip-tools==4.0.0
prompt-toolkit==2.0.9     # via ipython
psycopg2==2.8.3
ptyprocess==0.6.0         # via pexpect
pycparser==2.19           # via cffi
pygments==2.4.2           # via ipython
pyotp==2.3.0
pytz==2019.2              # via django
raven==6.10.0
redis==3.3.7              # via django-cacheops
requests==2.22.0
six==1.12.0               # via argon2-cffi, django-cacheops, django-extensions, pip-tools, prompt-toolkit, traitlets
sqlparse==0.3.0           # via django
traitlets==4.3.2          # via ipython
ua-parser==0.8.0
urllib3==1.25.3           # via requests
user-agents==2.0
uwsgi==2.0.18
wcwidth==0.1.7            # via prompt-toolkit
zxcvbn==4.4.28

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.2.0        # via ipdb, ipython
